requirements.txt
langchain
langchain-groq
langchain-community
youtube-transcript-api
pypdf
python-dotenv
beautifulsoup4
streamlit

